feat: admin site pagenation and delete

+++ backend/controller/site-controller.js
@@ -21,23 +21,47 @@ const siteController = {
       colorset,
       blocks,
     });
-    return res.status(201).json(newSite);
+    return res.ok(200, newSite);
   }),
+
   getSiteInfo: asyncHandler(async (req, res) => {
+    if (is.emptyObject(req.body)) {
+      throw new BadRequestError(
+        "headers의 Content-Type을 application/json으로 설정해주세요"
+      );
+    }
     const siteId = req.params.siteId;
     const site = await siteService.getSiteInfo(siteId);
-    res.status(200).json(site);
+    return res.ok(200, site);
   }),
+
   getSitesInfo: asyncHandler(async (req, res) => {
+    if (is.emptyObject(req.body)) {
+      throw new BadRequestError(
+        "headers의 Content-Type을 application/json으로 설정해주세요"
+      );
+    }
     const sites = await siteService.getSitesInfo();
-    res.status(200).json(sites);
+    return res.ok(200, sites);
   }),
+
   getUserSites: asyncHandler(async (req, res) => {
+    if (is.emptyObject(req.body)) {
+      throw new BadRequestError(
+        "headers의 Content-Type을 application/json으로 설정해주세요"
+      );
+    }
     const userId = req.params.userId;
     const sites = await siteService.getSites(userId);
-    res.status(200).json(sites);
+    return res.ok(200, sites);
   }),
+
   updateSite: asyncHandler(async (req, res) => {
+    if (is.emptyObject(req.body)) {
+      throw new BadRequestError(
+        "headers의 Content-Type을 application/json으로 설정해주세요"
+      );
+    }
     const siteId = req.params.siteId;
     const name = req.body.name;
     const domain = req.body.domain;
@@ -55,12 +79,18 @@ const siteController = {
       ...(blocks && { blocks }),
     };
     const updatedSiteInfo = await siteService.updateSite(siteId, toUpdate);
-    res.status(200).json(updatedSiteInfo);
+    return res.ok(200, updatedSiteInfo);
   }),
+
   deleteSite: asyncHandler(async (req, res) => {
+    if (is.emptyObject(req.body)) {
+      throw new BadRequestError(
+        "headers의 Content-Type을 application/json으로 설정해주세요"
+      );
+    }
     const siteId = req.params.siteId;
     const deletedSite = await siteService.deleteSiteBySiteId(siteId);
-    res.status(200).json(deletedSite);
+    return res.ok(200, deletedSite);
   }),
 };
